+++ crates/diffguard-types/tests/properties.rs
@@ -180,7 +180,7 @@ fn arb_diff_meta() -> impl Strategy<Value = DiffMeta> {
         arb_non_empty_string(), // head
         0u32..100,              // context_lines
         arb_scope(),            // scope
-        0u32..1000,             // files_scanned
+        0u64..1000,             // files_scanned (u64 to handle large repos)
         0u32..10000,            // lines_scanned
     )
         .prop_map(
@@ -1154,7 +1154,7 @@ mod unit_tests {
                 head: "HEAD".to_string(),
                 context_lines: u32::MAX,
                 scope: Scope::Added,
-                files_scanned: u32::MAX,
+                files_scanned: u64::MAX,
                 lines_scanned: u32::MAX,
             },
             findings: vec![Finding {
